add function ventoy_password_get
